Refactor python unittests
